Add files via upload
Predicting Housing Prices using Tensorflow + Cloud AI Platform
This notebook will show you how to create a tensorflow model, train it on the cloud in a distributed fashion across multiple CPUs or GPUs and finally deploy the model for online prediction. We will demonstrate this by building a model to predict housing prices.